Extract to constant time and days options
