+++ src/app/register/cooperativa/onboarding/components/FormStep3.jsx
@@ -3,17 +3,10 @@ import { useState } from 'react';
 import { Card } from '@nextui-org/react';
 import TimeSelector from '@/components/TimeSelector';
 import GreenRoundedButton from '@/components/greenRoundedButton';
+import { timeOptions, daysOptions } from '@constants/dateAndTime';
 
 const OnboardingCooperativaFormStep3 = ({prevStep, nextStep, setAvailableSchedule, availableSchedule}) => {
-  const [days, setDays] = useState(availableSchedule || [
-    { id: 1, name: 'MONDAY', name_spanish: 'Lunes', active: false, begin_at: "00:00", end_at: "00:00" },
-    { id: 2, name: 'TUESDAY', name_spanish: 'Martes', active: false, begin_at: "00:00", end_at: "00:00" },
-    { id: 3, name: 'WEDNESDAY', name_spanish: 'Miercoles', active: false, begin_at: "00:00", end_at: "00:00" },
-    { id: 4, name: 'THURSDAY', name_spanish: 'Jueves', active: false, begin_at: "00:00", end_at: "00:00" },
-    { id: 5, name: 'FRIDAY', name_spanish: 'Viernes', active: false, begin_at: "00:00", end_at: "00:00" },
-    { id: 6, name: 'SATURDAY', name_spanish: 'Sábado', active: false, begin_at: "00:00", end_at: "00:00" },
-    { id: 7, name: 'SUNDAY', name_spanish: 'Domingo', active: false, begin_at: "00:00", end_at: "00:00" }
-  ]);
+  const [days, setDays] = useState(availableSchedule || daysOptions);
 
   const toggleDay = (id) => {
     setDays(days.map(day => 
@@ -26,14 +19,6 @@ const OnboardingCooperativaFormStep3 = ({prevStep, nextStep, setAvailableSchedul
       day.id === id ? { ...day, [field]: value } : day
     ));
   };
-
-  const timeOptions = [
-    '00:00', '00:30', '01:00', '01:30', '02:00', '02:30', '03:00', '03:30', '04:00', '04:30', '05:00', '05:30', '06:00', '06:30',
-    '07:00', '07:30', '08:00', '08:30', '09:00', '09:30', '10:00', '10:30', '11:00', '11:30', '12:00', '12:30', '13:00', '13:30',
-    '14:00', '14:30', '15:00', '15:30', '16:00', '16:30', '17:00', '17:30', '18:00', '18:30', '19:00', '19:30', '20:00', '20:30',
-    '21:00', '21:30', '22:00', '22:30', '23:00', '23:30'
-  ];
-
 
   const nextForm = () => {
     setAvailableSchedule(days);
